Automatic profile setting for tuned 2.9
Tuned 2.9 introduces automatic and manual profile selection.
If there is a file /etc/tuned/profile_mode containing "manual"
on a system with tuned 2.9, without this PR the appropriate OCP
profiles will not be set.  This PR removes the file prior to
tuned restart, in addition to /etc/tuned/active_profile to
make the tuned profile selection automatic.