Do nothing if input is already a pairs string
The method `map_from_pairs` works if the input is already a dictionary.
This changes adds the same behavioru to the counter part `map_to_pairs`
method.

If the input is a string, this string will be returned unchanged.
Validation of the input realy is a key value pair is left to the
utilising code.