Fix OpenShift SDN/OVS pod restart during upgrades
When the OpenShift SDN/OVS DaemonSets were upgraded during control plane
upgrades with an updateStrategy of RollingUpdate, an upgrade of the pods
in the entire cluster was performed.  This caused unexpected network and
application outages on nodes.

This patch updates the following:
* Changed the updateStrategy for SDN/OVS pods to OnDelete in the template,
  affects new installs.
* Added control plane upgrade task to modify SDN/OVS daemonsets to use
  OnDelete updateStrategy
* Added node upgrade task to delete all SDN/OVS pods while nodes are drained

Network outages for nodes should only occur during the node upgrade when
nodes are drained.