Ensure that metrics node selectors are dicts (#16)
The replication controllers for cassandra, hawkular and heapster expect
a python dict containing the node selectors, but the respective tasks
provide a string. This is unfortunate because the check in the template
which decides if there are any selectors to apply returns true both in
the presence of a non-empty string and a non-empty dict.

This commit applies `map_from_pairs` to the configured node selector
variables to ensure that the templates are provided with the node
selectors in a dict.