Implement option for iptables-min-sync-period in role openshift_node
On one of our clusters we are investigating issues with slow updates of
changing endpoints for services. RedHat support suggested that we try
increasing 'iptables-min-sync-period' to give iptables more time to
catch up when many rules need to be changed frequently.

This commit modifies the role openshift_node to allow configuring
'iptables-min-sync-period'. The default value for
iptables-min-sync-period is set to the upstream default of 2 seconds.