Add readiness + liveness probes Service Catalog API server
The kubernetes garbage collector is removing secrets,
if the OwnerReference is not available during
the service catalog api server pods are starting up.

Because the service catalog pods are started from a
daemonset, the master drain do not evict them.
This causes the kubernetes controller and the
service catalog starting at the same time.
During the startup of the controller, the
garbage collector does run at the same time
the api service is starting up.