Fix openshift_logging_elasticsearch handler (Ansible 2.8)
In Ansible 2.7 the _cluster_component var was not evaluated in the task
name and was allowing the task to complete.  In Ansible 2.8, the var is
attempted to be evaluated but does not exist as it is assigned in the
loop_var of the task itself.  The _cluster_component var has been removed
from the task handler task name.