Ensure master facts are set executing node config
Running the openshift-node playbook can fail, when the facts
have peen purged. The variable openshift_node_master_api_url
is required by the openshift-node config roles, but not set
in this case.

To ensure the variable is going to be available, enforce
gathering the master facts as already done in the
openshift-node scalup playbook.