gluster: Heketi template does not contain LVM-wrapper on containerized deployment
Due to an incorrect statement in the heketi-template.yml.j2 template,
the HEKETI_LVMWRAPPER environment variable is only incluseded in
deployments where Gluster runs on an external cluster, and not in the
case where Gluster is deployed on OpenStack in containers.

The environment variable is needed to specify a wrapper for LVM commands
so that the commands can be run on the container node, outside the
container. This improved the stability of the LVM configuration.

Fixes: RHBZ#1791667
Reported-by: Aditya Ramteke <aramteke@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>