Configure always a docker-registry config
The config file was only configured, if
object storage has been used. Because
the docker client > v19.03.0 is going
to remove the support for v2 schema v1,
we need always to add a config file
enabling the option acceptschema2
by default as it is already for
object based backend configurations.