Bug 1867134: Remove log-openvswitch mount from ovs daemonset
In PR #11702 we moved the logging to be inside the container. We
no longer store logs on host. The only reason why we kept the
log-openvswitch mount from host to pod was to cleanup the old
log files. Since it has been over a year and there are some
weird permission issues faced in the bug revolving this mounting
it is safe to remove this logic in 3.11.

Note: We might need to add some documentation for the users that
might have residue log files on the host in order to remove them.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>